backend/requirements.txt
ruff==0.12.3
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
py-pg-notify==1.0.2
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic-settings==2.1.0
pytest==7.4.3
python-json-logger==2.0.7
gunicorn==21.2.0